refactor navigator data handling to naviContainer
